Upates Providers' register code.
A provider object can now be registered with the Providers object. It will
check whether it already has a matching Provider model object in the database
and associate it with that. It will update all it's associated URLs.
Otherwise, a new model provider will be instantiated.
Then, it will be added into by_url in the manner: {url: (in_db, provider)} for
each url.